move custom validators to validators.py
